change folder name to routers, move auth middleware to middleware folder. remove view engine, replace routes with json responses
