+++ src/routes/api/auth/logout.ts
