+++ src/routes/api/auth/signup.ts
