+++ src/routes/index.ts
@@ -17,27 +17,15 @@ router.use((req: Request, res: Response, next: NextFunction) => {
 
 router.use('/api/', api);
 
-router.get('/', (req: Request, res: Response) => {
-  res.render('index', {title: 'Express'});
-});
-
-// catch 404 and forward to error handler
 router.use((req: Request, res: Response, next: NextFunction) => {
-  // console.log(req);
-  // TODO: log req.baseUrl correct.
-  logger.info(`requested url: ${req.baseUrl}`);
-  // next(createError(404));
+  logger.info(`requested url: ${req.url}`);
+  res.status(404);
+  return res.json({success: false, error: {message: 'Not Found'}});
 });
 
-// error handler
 router.use((err: CustomErrback, req: Request, res: Response) => {
-  // set locals, only providing error in development
-  res.locals.message = err.message;
-  res.locals.error = req.app.get('env') === 'development' ? err : {};
-
-  // render the error page
   res.status(err.status || 500);
-  res.render('error');
+  return res.json({success: false, error: {message: req.app.get('env') === 'development' ? err.message : ''}});
 });
 
 export {
